+++ side_experiments/assume_result_w_obs/solution/absolute_return_node_helpers/explore_helpers.cpp
@@ -8,7 +8,7 @@ using namespace std;
 void AbsoluteReturnNode::explore_activate(Problem* problem,
 										  vector<ContextLayer>& context,
 										  RunHelper& run_helper) {
-	vector<double> world_location = problem_type->relative_to_world(
+	vector<int> world_location = problem_type->relative_to_world(
 		context.back().starting_location, this->location);
 	problem->return_to_location(world_location);
