+++ side_experiments/assume_result_w_obs/solution/return_node_helpers/explore_helpers.cpp
@@ -8,7 +8,7 @@ void ReturnNode::explore_activate(Problem* problem,
 								  vector<ContextLayer>& context,
 								  RunHelper& run_helper) {
 	if (this->previous_location != NULL) {
-		map<AbstractNode*, pair<vector<double>,vector<double>>>::iterator it
+		map<AbstractNode*, pair<vector<int>,vector<double>>>::iterator it
 			= context.back().node_history.find(this->previous_location);
 		if (it != context.back().node_history.end()) {
 			problem->return_to_location(it->second.first);
