+++ side_experiments/assume_result_w_obs/world/minesweeper.h
@@ -26,13 +26,13 @@ public:
 	Minesweeper();
 
 	void get_observations(std::vector<double>& obs,
-						  std::vector<std::vector<double>>& locations);
+						  std::vector<std::vector<int>>& locations);
 	void perform_action(Action action);
 	double score_result(int num_analyze,
 						int num_actions);
 
-	std::vector<double> get_location();
-	void return_to_location(std::vector<double>& location);
+	std::vector<int> get_location();
+	void return_to_location(std::vector<int>& location);
 
 	Problem* copy_and_reset();
 	Problem* copy_snapshot();
@@ -51,10 +51,10 @@ public:
 	int num_possible_actions();
 	Action random_action();
 
-	std::vector<double> relative_to_world(std::vector<double>& comparison,
-										  std::vector<double>& relative_location);
-	std::vector<double> world_to_relative(std::vector<double>& comparison,
-										  std::vector<double>& world_location);
+	std::vector<int> relative_to_world(std::vector<int>& comparison,
+									   std::vector<int>& relative_location);
+	std::vector<int> world_to_relative(std::vector<int>& comparison,
+									   std::vector<int>& world_location);
 };
 
 #endif /* MINESWEEPER_H */
