+++ side_experiments/assume_result_w_obs/world/problem.h
@@ -8,13 +8,13 @@ public:
 	virtual ~Problem() {};
 
 	virtual void get_observations(std::vector<double>& obs,
-								  std::vector<std::vector<double>>& locations) = 0;
+								  std::vector<std::vector<int>>& locations) = 0;
 	virtual void perform_action(Action action) = 0;
 	virtual double score_result(int num_analyze,
 								int num_actions) = 0;
 
-	virtual std::vector<double> get_location() = 0;
-	virtual void return_to_location(std::vector<double>& location) = 0;
+	virtual std::vector<int> get_location() = 0;
+	virtual void return_to_location(std::vector<int>& location) = 0;
 
 	virtual Problem* copy_and_reset() = 0;
 	virtual Problem* copy_snapshot() = 0;
@@ -32,10 +32,10 @@ public:
 	virtual int num_possible_actions() = 0;
 	virtual Action random_action() = 0;
 
-	virtual std::vector<double> relative_to_world(std::vector<double>& comparison,
-												  std::vector<double>& relative_location) = 0;
-	virtual std::vector<double> world_to_relative(std::vector<double>& comparison,
-												  std::vector<double>& world_location) = 0;
+	virtual std::vector<int> relative_to_world(std::vector<int>& comparison,
+											   std::vector<int>& relative_location) = 0;
+	virtual std::vector<int> world_to_relative(std::vector<int>& comparison,
+											   std::vector<int>& world_location) = 0;
 };
 
 #endif /* PROBLEM_H */
